requirements.txt
alembic==1.8.1
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
charset-normalizer==2.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
docopt==0.6.2
Fiona==1.8.21
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
geopandas==0.11.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.71
Mako==1.2.1
MarkupSafe==2.1.1
munch==2.5.0
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pipwin==0.5.2
psycopg2-binary==2.9.3
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pyproj==3.3.1
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
Shapely==1.8.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
waitress==2.1.2
Werkzeug==2.1.2
zipp==3.8.1
